+++ src/features/feed/components/FeedHomeList/style.ts
@@ -1,0 +1,19 @@
+import styled from '@emotion/styled';
+
+export const NoResult = styled.div`
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+`;
+
+export const ObservedBox = styled.div`
+  height: 10px;
+`;
+
+export const NoMoreResult = styled.div`
+  display: flex;
+  justify-content: center;
+  padding: 1rem 0;
+`;
