+++ src/features/feed/components/index.ts
@@ -2,3 +2,4 @@ export { default as FeedItem } from './FeedItem';
 export { default as FeedSelectBucket } from './FeedSelectBucket';
 export { default as FeedItemsDetail } from './FeedItemsDetail';
 export { default as FeedList } from './FeedList';
+export { default as FeedHomeList } from './FeedHomeList';
